# Speed up JSON chunker.